rm custom addon and user search fields on /admin landing page
