+++ src/olympia/zadmin/admin.py
@@ -1,7 +1,11 @@
-from django.contrib import admin
 from django.conf import settings
+from django.contrib import admin, auth
+from django.core.exceptions import PermissionDenied
+from django.shortcuts import redirect
 from django.utils.html import format_html
 from django.urls import reverse
+
+from olympia.accounts.utils import redirect_for_login
 
 from . import models
 
@@ -40,6 +44,21 @@ def related_single_content_link(obj, related_field):
         return ''
 
 
+# Hijack the admin's login to use our pages.
+def login(request):
+    # if the user has permission, just send them to the index page
+    if request.method == 'GET' and admin.site.has_permission(request):
+        next_path = request.GET.get(auth.REDIRECT_FIELD_NAME)
+        return redirect(next_path or 'admin:index')
+    # otherwise, they're logged in but don't have permission return a 403.
+    elif request.user.is_authenticated:
+        raise PermissionDenied
+    else:
+        return redirect_for_login(request)
+
+
 admin.site.register(models.Config)
 admin.site.disable_action('delete_selected')
 admin.site.site_url = settings.EXTERNAL_SITE_URL
+admin.site.site_header = admin.site.index_title = 'AMO Administration'
+admin.site.login = login
