--- src/olympia/zadmin/templates/zadmin/addon-search.html
@@ -1,25 +1,0 @@
-{% extends "zadmin/base.html" %}
-
-{% block title %}{{ page_title('Add-on Search') }}{% endblock %}
-
-{% block content %}
-<h2>Search Add-ons</h2>
-{% include "zadmin/includes/search_addons.html" %}
-
-{% if q %}
-<h3>Results for <b>{{ q }}</b> (1&ndash;{{ addons|length }} of {{ addons.count() }}):</h3>
-  {% if addons %}
-  <ul>
-    {% for addon in addons %}
-    <li>
-      <a href="{{ url('admin:addons_addon_change', addon.id) }}">
-        {{ addon.name }} [{{ addon.id }}]</a>
-    </li>
-    {% endfor %}
-  </ul>
-  {% else %}
-    <p>Nothing found!</p>
-  {% endif %}
-{% endif %}
-
-{% endblock %}
