--- src/olympia/zadmin/templates/zadmin/includes/search_users.html
@@ -1,7 +1,0 @@
-<form method="get" action="{{ url('users.admin_edit', '{id}') }}"
-      data-search-url="{{ url('zadmin.search', 'users', 'userprofile') }}" data-no-csrf>
-    <div>
-        <input class="searchbar" type="text" value="" name="q" placeholder="id, username, or email">
-        <p class="help">Enter the beginning of user's id, username, or email and select the user from the dropdown.</p>
-    </div>
-</form>
