+++ src/olympia/zadmin/templates/zadmin/index.html
@@ -5,12 +5,6 @@
 {% block content %}
 
 <div class="content">
-    <section>
-        <h2>Search Add-ons</h2>
-        {% include "zadmin/includes/search_addons.html" %}
-        <h2>Search Users</h2>
-        {% include "zadmin/includes/search_users.html" %}
-    </section>
     <aside>
         <h2>Admin Log</h2>
         {% if log %}
